test.html

<!DOCTYPE html>
<html lang="en">
<head>
  <meta charset="utf-8" />
  <title>Mocha</title>
  <meta name="viewport" content="width=device-width, initial-scale=1.0" />

  <script src="testharness/testharness.js" crossorigin="anonymous"></script>
  <script src="testharness/testharnessreport.js" crossorigin="anonymous"></script>
</head>
<body>
  <script type='module'>
    import {loadProgram, load} from "./sf2-service/read.js";

    promise_test(async () => {
      const sf2 = await load("file.sf2");
      const {shdrMap, zMap, preload, filterKV} = loadProgram(sf2, 0, 0);
      assert_greater_than(zMap.length, 0, "number of zones");
      assert_greater_than(Object.keys(shdrMap).length, 0, "number of zones");
    }, "loading program returns more than 0 zones");

    promise_test(async () => {
      const sf2 = await load("file.sf2");
      const {shdrMap, zMap, preload, filterKV} = loadProgram(sf2, 0, 0);
      await preload();
      console.log(zMap);
      assert_true(zMap[0].shdr.pcm instanceof Float32Array, "pcm data can be accessed synchronously after preload");
    }, "preload pcm data");

    promise_test(async () => {
      const times = [];
      const el = document.createElement('div');
      document.body.append(el);
      const sf2 = await load("file.sf2", {
        onHeader(type, bid, str) {
          el.append(new Text("program" + str))
          times.push(performance.now());
        }
      });
      assert_true(document.body.innerHTML.includes("program"));
      el.remove();
    }, "stream ui load as file is being parsed");
  </script>
  <script type='module'>
    import {test} from './tests/dropship.test.js';
    test();
  </script>
</body>
</html>
